+++ blob/lib/Pipeline.ts
@@ -1,9 +1,9 @@
 import {
   BaseRequestPolicy,
-  HttpClient as IHTTPClient,
+  HttpClient as IHttpClient,
   HttpHeaders,
   HttpOperationResponse,
-  HttpPipelineLogger as IHTTPPipelineLogger,
+  HttpPipelineLogger as IHttpPipelineLogger,
   HttpPipelineLogLevel,
   RequestPolicy,
   RequestPolicyFactory,
@@ -15,8 +15,8 @@ import {
 // Export following interfaces and types for customers who want to implement their
 // own RequestPolicy or HTTPClient
 export {
-  IHTTPClient,
-  IHTTPPipelineLogger,
+  IHttpClient,
+  IHttpPipelineLogger,
   HttpHeaders,
   HttpPipelineLogLevel,
   HttpOperationResponse,
@@ -34,8 +34,8 @@ export {
  * @interface IPipelineOptions
  */
 export interface IPipelineOptions {
-  logger?: IHTTPPipelineLogger;
-  HTTPClient?: IHTTPClient;
+  logger?: IHttpPipelineLogger;
+  HTTPClient?: IHttpClient;
 }
 
 /**
@@ -53,7 +53,7 @@ export class Pipeline {
   public readonly options: IPipelineOptions;
 
   /**
-   * Creates an instance of Pipeline. Customize HTTPClient by implementing IHTTPClient interface.
+   * Creates an instance of Pipeline. Customize HTTPClient by implementing IHttpClient interface.
    *
    * @param {RequestPolicyFactory[]} factories
    * @param {IPipelineOptions} [options={}]